+++ gui.cpccImageMac.h
@@ -184,8 +184,8 @@ public: /// ctors
 	
     
 public: /// functions
-	virtual const int getWidth(void) const	override {   return bmpPtr ? bmpPtr.pixelsWide : 0;  }
-	virtual const int getHeight(void) const	override {	return bmpPtr ? bmpPtr.pixelsHigh : 0;  }
+	virtual const int getWidth(void) const	override {   return bmpPtr ? (int) bmpPtr.pixelsWide : 0;  }
+	virtual const int getHeight(void) const	override {	return bmpPtr ? (int) bmpPtr.pixelsHigh : 0;  }
     
     virtual void initWithSizeAndColor(const int aWidth, const int aHeight, const cpccColor &aColor) override
     {
